Tests start their own server for tests
I do need to have a tmux session called 'goo_server' fired up for this
to work.

It's jenky for sure, but I can't figure out how to get the test file to
fire up its own tmux session, detach from it and send a request.  The
commands seem to work from the command line, but not the test file.