Fix content length bug
Because I was calling a methods to assemble the 'page' and it was
mutating the body array when I did that, the diagnostic information was
getting called repeatedly.  It juuuust started to show double on the
bottom of the screen.

By changing the method name to 'assemble_page' and using an accessor for
the page, it made it more clear what was going on and gave me more
control over the body of the page.